+++ lib/src/engine/transport/transport.dart
@@ -117,8 +117,7 @@ abstract class Transport extends EventEmitter {
   /// @param {String} data
   /// @api private
   void onData(data) {
-    var packet = PacketParser.decodePacket(data,
-        binaryType: socket?.binaryType, utf8decode: false);
+    var packet = PacketParser.decodePacket(data, socket!.binaryType);
     onPacket(packet);
   }