+++ test/io_client.dart
@@ -5,18 +5,24 @@ import 'dart:async';
 import 'package:socket_io_client/socket_io_client.dart' as io;
 
 void main() {
-  var socket = io.io('http://localhost:3000', <String, dynamic>{
-    'transports': ['websocket'],
-    'extraHeaders': {'foo': 'bar'}
-  });
-  socket.on('connect', (_) {
-    print('connect');
-    socket.emit('msg', 'init');
+  var socket = io.io(
+      'http://localhost:3000',
+      io.OptionBuilder()
+          .setTransports(['websocket'])
+          // .disableAutoConnect()
+          .build());
+
+  // socket.connect();
+
+  socket.onConnect((_) {
+    socket.emit('toServer', 'init');
+
     var count = 0;
     Timer.periodic(const Duration(seconds: 1), (Timer countDownTimer) {
-      socket.emit('msg', count++);
+      socket.emit('toServer', count++);
     });
   });
+
   socket.on('event', (data) => print(data));
   socket.on('disconnect', (_) => print('disconnect'));
   socket.on('fromServer', (_) => print(_));