+++ test_nodejs/server.js
@@ -11,21 +11,16 @@
  * Copyright (C) 2019 Potix Corporation. All Rights Reserved.
  */
 'use strict';
-var http = require('http').createServer();
-var io = require('socket.io')(http);
-
-io.on('connection', function(socket){
-    console.log('connection default namespace');
-    socket.on('msg', function (data, ack) {
-        if (ack != null) {
-            ack(1, 2, 3);
-        }
-        console.log(`data from default => ${data}`);
-        socket.emit('fromServer', `${data}`);
+const app = require('express')();
+const server = require('http').createServer(app);
+const io = require('socket.io')(server);
+io.on('connection', userSocket => {
+    console.log('connected');
+    userSocket.on('toServer', data => {
+        userSocket.emit('fromServer', data);
     });
 });
-
-http.listen(3000, function(){
+server.listen(3000, function(){
     console.log('listening on *:3000');
 });